refactor: use js language files in place of json ones
