+++ src/i18n/DopplerIntlProvider.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { addLocaleData, IntlProvider } from 'react-intl';
 import en from 'react-intl/locale-data/en';
 import es from 'react-intl/locale-data/es';
-import messages_es from './es.json';
-import messages_en from './en.json';
+import messages_es from './es';
+import messages_en from './en';
 import { flattenMessages } from './utils';
 
 const messages = {
